refactor: DTO structure
