+++ src/main/java/com/future/restoapp/controller/DefaultExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.future.restoapp.controller;
 
-import com.future.restoapp.model.dto.ErrorResponse;
+import com.future.restoapp.model.dto.core.ErrorResponse;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
