+++ src/main/java/com/future/restoapp/controller/MenuController.java
@@ -1,10 +1,9 @@
 package com.future.restoapp.controller;
 
 import com.future.restoapp.controller.path.MenuControllerPath;
-import com.future.restoapp.controller.path.UserControllerPath;
-import com.future.restoapp.model.dto.MenuCreateRequest;
-import com.future.restoapp.model.dto.MenuUpdateRequest;
-import com.future.restoapp.model.dto.SuccessResponse;
+import com.future.restoapp.model.dto.menu.MenuCreateRequest;
+import com.future.restoapp.model.dto.menu.MenuUpdateRequest;
+import com.future.restoapp.model.dto.core.SuccessResponse;
 import com.future.restoapp.model.entity.Menu;
 import com.future.restoapp.model.entity.Menu.MenuCategory;
 import com.future.restoapp.service.AssetService;
