+++ src/main/java/com/future/restoapp/controller/UserController.java
@@ -1,10 +1,10 @@
 package com.future.restoapp.controller;
 
 import com.future.restoapp.controller.path.UserControllerPath;
-import com.future.restoapp.model.dto.RegisterRequest;
-import com.future.restoapp.model.dto.SuccessResponse;
-import com.future.restoapp.model.dto.UserResponse;
-import com.future.restoapp.model.dto.UserUpdateRequest;
+import com.future.restoapp.model.dto.user.RegisterRequest;
+import com.future.restoapp.model.dto.core.SuccessResponse;
+import com.future.restoapp.model.dto.user.UserResponse;
+import com.future.restoapp.model.dto.user.UserUpdateRequest;
 import com.future.restoapp.model.entity.User;
 import com.future.restoapp.service.UserService;
 import io.swagger.v3.oas.annotations.tags.Tag;
