+++ src/main/java/com/future/restoapp/model/dto/menu/MenuUpdateRequest.java
@@ -1,16 +1,17 @@
-package com.future.restoapp.model.dto;
+package com.future.restoapp.model.dto.menu;
+
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.future.restoapp.model.entity.Reservation;
+import com.future.restoapp.model.entity.Menu;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.BeanUtils;
 
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
-import java.io.Serializable;
 import java.util.Optional;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -18,20 +19,22 @@ import java.util.Optional;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReservationAdminUpdateRequest implements Serializable {
+public class MenuUpdateRequest {
 
     @NotNull(message = "id must be specified")
-    @Positive(message = "id value must be positive")
+    @Positive(message = "id must be positive value")
     private Long id;
 
-    @NotNull(message = "isAccepted must be specified")
-    private Boolean isAccepted;
+    @Positive(message = "price must be positive value")
+    private Float price;
 
-    public Reservation toReservation(){
+    private Boolean isSold;
+
+    public Menu toMenu(){
         return Optional.of(this).map(dto -> {
-            Reservation reservation = Reservation.builder().build();
-            BeanUtils.copyProperties(dto, reservation);
-            return reservation;
+            Menu menu = Menu.builder().build();
+            BeanUtils.copyProperties(dto, menu);
+            return menu;
         }).orElse(null);
     }
