+++ src/test/java/com/future/restoapp/unitTest/controller/UserControllerTest.java
@@ -2,7 +2,7 @@ package com.future.restoapp.unitTest.controller;
 
 import com.future.restoapp.controller.UserController;
 import com.future.restoapp.controller.path.UserControllerPath;
-import com.future.restoapp.model.dto.RegisterRequest;
+import com.future.restoapp.model.dto.user.RegisterRequest;
 import com.future.restoapp.model.entity.User;
 import com.future.restoapp.service.UserService;
 import org.junit.jupiter.api.AfterEach;
